Move clients
